requirements.txt
inflect
numpy==1.23.1
pycocotools==2.0.7
mdit_py_plugins==0.3.3
linkify-it-py==2.0.2
MarkupSafe==2.1.3
contourpy==1.1.1
matplotlib==3.5.1
cycler==0.12.0
fonttools==4.42.1
kiwisolver==1.4.5
packaging==21.3
scikit-image==0.17.2
Pillow==8.4.0
pudb==2019.2
albumentations==1.3.0
opencv-python==4.1.2.30
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.5.0
omegaconf==2.1.1
test-tube==0.7.5
streamlit==1.27.2
einops==0.3.0
transformers==4.32.1
kornia==0.7.0
torchmetrics==0.7.3
yapf==0.32.0
safetensors==0.4.0